Move ObservationListRows component
